requirements.txt
streamlit==1.29.0
pyshorteners==1.0.1
pyyaml==6.0.1
networkx==3.2.1
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
typing-extensions==4.9.0
aiofiles==23.2.1
asyncio==3.4.3
reportlab==4.0.8
Pillow==10.1.0
